Оптимизировал компонент Header, разнес по папкам
